Pushing to github for future use
This project has been sitting on my hard drive and I don't want to lose
it when reformatting, so it makes sense to push it to my Github account.

This project uses graph cuts over a Markov random field formulation of
an image to perform binary image segmentation.  See the README.txt and
the accompanying pdf for an explanation and instructions on how to run.